organized packages for AVE and Stats screens
