+++ inGroove/app/src/main/Java/com/cmput301f17t07/ingroove/HabitStats/HabitStatsActivity.java
@@ -1,7 +1,9 @@
-package com.cmput301f17t07.ingroove;
+package com.cmput301f17t07.ingroove.HabitStats;
 
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.cmput301f17t07.ingroove.R;
 
 public class HabitStatsActivity extends AppCompatActivity {
