+++ inGroove/app/src/main/Java/com/cmput301f17t07/ingroove/HabitStats/HabitStatsContract.java
@@ -1,0 +1,16 @@
+package com.cmput301f17t07.ingroove.HabitStats;
+
+/**
+ * Created by fraserbulbuc on 2017-10-24.
+ */
+
+public interface HabitStatsContract {
+
+    public interface Viewable {
+
+    }
+
+    public interface Controlling {
+
+    }
+}
