+++ inGroove/app/src/main/Java/com/cmput301f17t07/ingroove/HabitStats/HabitStatsController.java
@@ -1,0 +1,8 @@
+package com.cmput301f17t07.ingroove.HabitStats;
+
+/**
+ * Created by fraserbulbuc on 2017-10-24.
+ */
+
+public class HabitStatsController {
+}
