+++ inGroove/app/src/main/Java/com/cmput301f17t07/ingroove/avehabit/AddViewEditHabitActivity.java
@@ -1,9 +1,11 @@
-package com.cmput301f17t07.ingroove;
+package com.cmput301f17t07.ingroove.avehabit;
 
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import com.cmput301f17t07.ingroove.R;
 
 public class AddViewEditHabitActivity extends AppCompatActivity {
 
@@ -12,4 +14,8 @@ public class AddViewEditHabitActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_add_view_edit_habit);
     }
+
+
+
+
 }
