+++ inGroove/app/src/main/Java/com/cmput301f17t07/ingroove/avehabit/AddViewEditHabitContract.java
@@ -1,0 +1,18 @@
+package com.cmput301f17t07.ingroove.avehabit;
+
+/**
+ * Created by fraserbulbuc on 2017-10-24.
+ */
+
+public interface AddViewEditHabitContract {
+
+    public interface Viewable {
+
+    }
+
+
+    public interface Controlling {
+
+    }
+
+}
