+++ inGroove/app/src/main/res/layout/activity_add_view_edit_habit.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t07.ingroove.AddViewEditHabitActivity">
+    tools:context="com.cmput301f17t07.ingroove.avehabit.AddViewEditHabitActivity">
 
 </android.support.constraint.ConstraintLayout>
