+++ inGroove/app/src/main/res/layout/activity_habit_stats.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.cmput301f17t07.ingroove.HabitStatsActivity">
+    tools:context="com.cmput301f17t07.ingroove.HabitStats.HabitStatsActivity">
 
 </android.support.constraint.ConstraintLayout>
